requirements.txt
python-dotenv~=1.0.1
colorama~=0.4.6
python-slugify~=8.0.3
EbookLib~=0.18
ao3-api~=2.3.0